customer_360/kafka_consumer/pyproject.toml
[project]
name = "kafka-validator"
version = "0.1.0"
description = "Generic Kafka consumer with Avro schema validation for Flink tables"
requires-python = ">=3.9"
dependencies = [
    "confluent-kafka[avro,schema-registry]>=2.3.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.7",
]

[project.scripts]
validate-records = "kafka_validator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
